refactor: split run controller phases
